AutoLayout修正, クラッシュポイント修正, ユーザーアイコンのキャッシュ化
