+++ Qiita_App/Qiita_App/UIImage+Extension.swift
@@ -23,22 +23,3 @@ extension UIImage {
         return newImage
     }
 }
-
-extension UIImageView {
-
-    func setImageByDefault(with url: URL) {
-
-        URLSession.shared.dataTask(with: url) { [weak self] data, response, error in
-            // Success
-            if error == nil, case .some(let result) = data, let image = UIImage(data: result) {
-                self?.image = image
-
-            // Failure
-            } else {
-                // error handling
-
-            }
-        }.resume()
-    }
-
-}
